+++ Synthlab/src/synthlab/internal/BasicPort.java
@@ -19,30 +19,30 @@ public class BasicPort implements Port
   }
 
   @Override
-  public void setValue(Integer value)
+  public void setValue(double value)
   {
     value_ = value;
   }
 
   @Override
-  public Integer getValue()
+  public double getValue()
   {
     return value_;
   }
-  
+
   @Override
   public Module getModule()
   {
     return module_;
   }
-  
+
   @Override
   public void setModule(Module module)
   {
     module_ = module;
   }
 
-  private String  name_;
-  private Integer value_;
-  private Module  module_;
+  private String name_;
+  private double value_;
+  private Module module_;
 }
